% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{aggregate_allele_counts}
\alias{aggregate_allele_counts}
\title{Aggregate allele counts across multiple MCMC samples}
\usage{
aggregate_allele_counts(allele_counts_list)
}
\description{
Aggregate allele counts across multiple MCMC samples
}
\keyword{internal}
