% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial_initialize_state}
\alias{binomial_initialize_state}
\title{Initialize state for binomial model}
\usage{
binomial_initialize_state(model_obj, threshold = 0.001)
}
\arguments{
\item{model_obj}{Model object}

\item{threshold}{Threshold for identifying single infections}
}
\value{
Initialized state
}
\description{
Initialize state for binomial model
}
\keyword{internal}
