% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial_loglikelihood_matrix}
\alias{binomial_loglikelihood_matrix}
\title{Binomial Model for SNP-Slice}
\usage{
binomial_loglikelihood_matrix(A, D, model_obj)
}
\arguments{
\item{A}{Allocation matrix}

\item{D}{Dictionary matrix}

\item{model_obj}{Model object containing data}
}
\value{
Log-likelihood value
}
\description{
Implementation of the binomial observation model for SNP-Slice.
This model assumes Y ~ Binomial(R, proportions).
Log-likelihood for binomial model (matrix version)
}
\keyword{internal}
