% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial_resolve_exceptions}
\alias{binomial_resolve_exceptions}
\title{Resolve exceptions for binomial model}
\usage{
binomial_resolve_exceptions(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Resolve exceptions for binomial model
}
\keyword{internal}
