% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{build_multinomial_model}
\alias{build_multinomial_model}
\title{Build the multinomial model object}
\usage{
build_multinomial_model(processed_data, alpha, dict_prior = "empirical")
}
\arguments{
\item{processed_data}{Output of \code{\link{load_dataframe_multiallelic}} or
the list-input path of \code{\link{preprocess_data}} with
\code{model = "multinomial"}.}

\item{alpha}{IBP concentration parameter}

\item{dict_prior}{See \code{\link{resolve_dict_prior}}.}
}
\value{
Model object list (class assigned by \code{\link{create_model}}).
}
\description{
Build the multinomial model object
}
\keyword{internal}
