% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_initialize_state}
\alias{categorical_initialize_state}
\title{Initialize state for categorical model}
\usage{
categorical_initialize_state(model_obj, threshold = 0.001)
}
\arguments{
\item{model_obj}{Model object}

\item{threshold}{Threshold for identifying single infections}
}
\value{
Initialized state
}
\description{
Initialize state for categorical model
}
\keyword{internal}
