% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_loglikelihood_matrix}
\alias{categorical_loglikelihood_matrix}
\title{Log-likelihood for categorical model (matrix version)}
\usage{
categorical_loglikelihood_matrix(A, D, model_obj)
}
\arguments{
\item{A}{Allocation matrix}

\item{D}{Dictionary matrix}

\item{model_obj}{Model object containing data}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood for categorical model (matrix version)
}
\keyword{internal}
