% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_prop_bin}
\alias{categorical_prop_bin}
\title{Map proportion to lookup-table row (1-based)}
\usage{
categorical_prop_bin(prop)
}
\description{
Map proportion to lookup-table row (1-based)
}
\keyword{internal}
