% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_resolve_exceptions}
\alias{categorical_resolve_exceptions}
\title{Resolve exceptions for categorical model}
\usage{
categorical_resolve_exceptions(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Resolve exceptions for categorical model
}
\keyword{internal}
