% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_y_bin}
\alias{categorical_y_bin}
\title{Map observation to lookup-table column (1-based)}
\usage{
categorical_y_bin(y)
}
\description{
Map observation to lookup-table column (1-based)
}
\keyword{internal}
