% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{create_results_object}
\alias{create_results_object}
\title{Create results object}
\usage{
create_results_object(mcmc_result, model_obj, processed_data)
}
\arguments{
\item{mcmc_result}{MCMC results}

\item{model_obj}{Model object}

\item{processed_data}{Processed data}
}
\value{
snp_slice_results object
}
\description{
Create results object
}
\keyword{internal}
