% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{dictionary_logprior}
\alias{dictionary_logprior}
\title{Log prior of the dictionary under the model's prior}
\usage{
dictionary_logprior(D, model_obj)
}
\arguments{
\item{D}{Dictionary matrix}

\item{model_obj}{Model object}
}
\value{
Log prior probability
}
\description{
Dispatches to \code{model_obj$logprior_d} when the model defines one and to
the Bernoulli(\code{rho}) prior in \code{\link{logprior_d}} otherwise.
}
\keyword{internal}
