% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\name{end_timer}
\alias{end_timer}
\title{End timing a function}
\usage{
end_timer(function_name)
}
\arguments{
\item{function_name}{Name of the function being timed}
}
\description{
End timing a function
}
\keyword{internal}
