% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\docType{data}
\name{example_read0}
\alias{example_read0}
\title{Example Read Count Data - Reference Alleles}
\format{
A matrix of reference allele read counts with 200 hosts and 96 SNPs
}
\source{
Simulated data for package testing
}
\usage{
example_read0
}
\description{
Example reference allele read count data for testing and demonstration purposes.
}
\keyword{datasets}
