% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\docType{data}
\name{example_read1}
\alias{example_read1}
\title{Example Read Count Data - Alternate Alleles}
\format{
A matrix of alternate allele read counts with 200 hosts and 96 SNPs
}
\source{
Simulated data for package testing
}
\usage{
example_read1
}
\description{
Example alternate allele read count data for testing and demonstration purposes.
}
\keyword{datasets}
