% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\docType{data}
\name{example_snp_data}
\alias{example_snp_data}
\title{Example SNP Data}
\format{
A list containing:
\describe{
\item{read0}{Matrix of reference allele read counts}
\item{read1}{Matrix of alternate allele read counts}
}
}
\source{
Simulated data for package testing
}
\usage{
example_snp_data
}
\description{
Example SNP data for testing and demonstration purposes.
}
\keyword{datasets}
