% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_kstar}
\alias{get_kstar}
\title{Get last active feature index}
\usage{
get_kstar(A)
}
\arguments{
\item{A}{Allocation matrix}
}
\value{
Index of last active feature
}
\description{
Get last active feature index
}
\keyword{internal}
