% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_m}
\alias{get_m}
\title{Get column sums of allocation matrix}
\usage{
get_m(A)
}
\arguments{
\item{A}{Allocation matrix}
}
\value{
Column sums
}
\description{
Get column sums of allocation matrix
}
\keyword{internal}
