% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{get_mhratio}
\alias{get_mhratio}
\title{Metropolis-Hastings acceptance ratio}
\usage{
get_mhratio(logp1, logp0)
}
\arguments{
\item{logp1}{Log probability of state 1}

\item{logp0}{Log probability of state 0}
}
\value{
Acceptance probability
}
\description{
Metropolis-Hastings acceptance ratio
}
\keyword{internal}
