% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_mustar}
\alias{get_mustar}
\title{Get mu value for last active feature}
\usage{
get_mustar(A, mu)
}
\arguments{
\item{A}{Allocation matrix}

\item{mu}{Stick-breaking weights}
}
\value{
Mu value for last active feature
}
\description{
Get mu value for last active feature
}
\keyword{internal}
