% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\name{get_performance_summary}
\alias{get_performance_summary}
\title{Get performance summary}
\usage{
get_performance_summary()
}
\value{
Performance summary data frame
}
\description{
Get performance summary
}
