% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator (infix)}
\usage{
x \%||\% default
}
\arguments{
\item{x}{Value to check}

\item{default}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise default
}
\description{
Null coalescing operator (infix)
}
\keyword{internal}
