% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{gridsample}
\alias{gridsample}
\title{Grid sampling for constrained updates}
\usage{
gridsample(lb, ub, logf, ...)
}
\arguments{
\item{lb}{Lower bound}

\item{ub}{Upper bound}

\item{logf}{Log density function}

\item{...}{Additional arguments for logf}
}
\value{
Sampled value
}
\description{
Grid sampling for constrained updates
}
\keyword{internal}
