% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\name{init_performance_tracking}
\alias{init_performance_tracking}
\title{Initialize performance tracking}
\usage{
init_performance_tracking()
}
\description{
Initialize performance tracking
}
\keyword{internal}
