% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_cpp_obs_args}
\alias{kernel_cpp_obs_args}
\title{Obs cache for a compiled kernel call (kernel object or model)}
\usage{
kernel_cpp_obs_args(kernel, model_obj)
}
\description{
Obs cache for a compiled kernel call (kernel object or model)
}
\keyword{internal}
