% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_llik_tab}
\alias{kernel_llik_tab}
\title{Lookup table passed to compiled categorical updates}
\usage{
kernel_llik_tab(model_obj)
}
\description{
Lookup table passed to compiled categorical updates
}
\keyword{internal}
