% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_model_uses_cpp_cache}
\alias{kernel_model_uses_cpp_cache}
\title{Whether a model can use the compiled kernel obs cache}
\usage{
kernel_model_uses_cpp_cache(model_obj)
}
\description{
Whether a model can use the compiled kernel obs cache
}
\keyword{internal}
