% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_obs_args}
\alias{kernel_obs_args}
\title{Arguments for cached obs layout passed into cpp kernel entry points}
\usage{
kernel_obs_args(model_obj)
}
\description{
Arguments for cached obs layout passed into cpp kernel entry points
}
\keyword{internal}
