% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_obs_cache_ready}
\alias{kernel_obs_cache_ready}
\title{Whether precomputed kernel obs cache is attached to a model}
\usage{
kernel_obs_cache_ready(model_obj)
}
\description{
Whether precomputed kernel obs cache is attached to a model
}
\keyword{internal}
