% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{kernel_with_obs_cache}
\alias{kernel_with_obs_cache}
\title{Attach obs cache vectors to a compiled kernel adapter list}
\usage{
kernel_with_obs_cache(kernel, model_obj)
}
\description{
Attach obs cache vectors to a compiled kernel adapter list
}
\keyword{internal}
