% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{logf_newfeature}
\alias{logf_newfeature}
\title{Log density for new features}
\usage{
logf_newfeature(x, N, alpha)
}
\arguments{
\item{x}{Value}

\item{N}{Number of hosts}

\item{alpha}{IBP concentration parameter}
}
\value{
Log density
}
\description{
Log density for new features
}
\keyword{internal}
