% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{logf_oldfeature}
\alias{logf_oldfeature}
\title{Log density for existing features}
\usage{
logf_oldfeature(x, m, N)
}
\arguments{
\item{x}{Value}

\item{m}{Feature count}

\item{N}{Number of hosts}
}
\value{
Log density
}
\description{
Log density for existing features
}
\keyword{internal}
