% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{logprior_a}
\alias{logprior_a}
\title{Log prior for allocation matrix A}
\usage{
logprior_a(A, mu, alpha)
}
\arguments{
\item{A}{Allocation matrix}

\item{mu}{Stick-breaking weights}

\item{alpha}{IBP concentration parameter}
}
\value{
Log prior probability
}
\description{
Log prior for allocation matrix A
}
\keyword{internal}
