% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{logprior_d}
\alias{logprior_d}
\title{Log prior for dictionary matrix D}
\usage{
logprior_d(D, rho)
}
\arguments{
\item{D}{Dictionary matrix}

\item{rho}{Dictionary sparsity parameter}
}
\value{
Log prior probability
}
\description{
Log prior for dictionary matrix D
}
\keyword{internal}
