% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{logprior_mu}
\alias{logprior_mu}
\title{Log prior for stick-breaking weights mu}
\usage{
logprior_mu(mu, alpha, N)
}
\arguments{
\item{mu}{Stick-breaking weights}

\item{alpha}{IBP concentration parameter}

\item{N}{Number of hosts}
}
\value{
Log prior probability
}
\description{
Log prior for stick-breaking weights mu
}
\keyword{internal}
