% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{map_estimates}
\alias{map_estimates}
\title{Extract MAP allocation and dictionary matrices from a chain state}
\usage{
map_estimates(map_state)
}
\arguments{
\item{map_state}{MAP state of a chain}
}
\value{
List with \code{allocation_matrix} and \code{dictionary_matrix},
restricted to strains with at least one host
}
\description{
Extract MAP allocation and dictionary matrices from a chain state
}
\keyword{internal}
