% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{model_type_id}
\alias{model_type_id}
\title{Map observation model name to compiled kernel type id}
\usage{
model_type_id(model_name)
}
\description{
Map observation model name to compiled kernel type id
}
\keyword{internal}
