% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{multinomial_loglik_cpp}
\alias{multinomial_loglik_cpp}
\title{Compiled total log-likelihood for the multinomial model}
\usage{
multinomial_loglik_cpp(state, model_obj)
}
\description{
Compiled total log-likelihood for the multinomial model
}
\keyword{internal}
