% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_kernel.R
\name{multinomial_update_iter_cpp}
\alias{multinomial_update_iter_cpp}
\title{Fused compiled slice iteration for the multinomial model}
\usage{
multinomial_update_iter_cpp(state, model_obj)
}
\description{
Fused compiled slice iteration for the multinomial model
}
\keyword{internal}
