% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{negative_binomial_initialize_state}
\alias{negative_binomial_initialize_state}
\title{Initialize state for negative binomial model}
\usage{
negative_binomial_initialize_state(model_obj, threshold = 0.001)
}
\arguments{
\item{model_obj}{Model object}

\item{threshold}{Threshold for identifying single infections}
}
\value{
Initialized state
}
\description{
Initialize state for negative binomial model
}
\keyword{internal}
