% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{negative_binomial_loglikelihood_matrix}
\alias{negative_binomial_loglikelihood_matrix}
\title{Negative Binomial Model for SNP-Slice}
\usage{
negative_binomial_loglikelihood_matrix(A, D, model_obj)
}
\arguments{
\item{A}{Allocation matrix}

\item{D}{Dictionary matrix}

\item{model_obj}{Model object containing data}
}
\value{
Log-likelihood value
}
\description{
Implementation of the negative binomial observation model for SNP-Slice.
This is the recommended default model for most applications.
Log-likelihood for negative binomial model (matrix version)
}
\keyword{internal}
