% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{negative_binomial_loglikelihood_vector}
\alias{negative_binomial_loglikelihood_vector}
\title{Log-likelihood for negative binomial model (vector version)}
\usage{
negative_binomial_loglikelihood_vector(propvec, yvec, rvec)
}
\arguments{
\item{propvec}{Proportion vector}

\item{yvec}{Observed counts vector}

\item{rvec}{Total counts vector}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood for negative binomial model (vector version)
}
\keyword{internal}
