% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{negative_binomial_resolve_exceptions}
\alias{negative_binomial_resolve_exceptions}
\title{Resolve exceptions for negative binomial model}
\usage{
negative_binomial_resolve_exceptions(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Resolve exceptions for negative binomial model
}
\keyword{internal}
