% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{or_null}
\alias{or_null}
\title{Helper function for default values}
\usage{
or_null(x, default)
}
\arguments{
\item{x}{Value to check}

\item{default}{Default value}
}
\value{
x if not NULL, otherwise default
}
\description{
Helper function for default values
}
\keyword{internal}
