% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\name{performance_env}
\alias{performance_env}
\title{Performance Logger for SNP-Slice}
\usage{
performance_env
}
\description{
Utility functions for tracking performance and identifying bottlenecks
}
