% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_loglikelihood_matrix}
\alias{poisson_loglikelihood_matrix}
\title{Poisson Model for SNP-Slice}
\usage{
poisson_loglikelihood_matrix(A, D, model_obj)
}
\arguments{
\item{A}{Allocation matrix}

\item{D}{Dictionary matrix}

\item{model_obj}{Model object containing data}
}
\value{
Log-likelihood value
}
\description{
Implementation of the Poisson observation model for SNP-Slice.
This model assumes Y ~ Poisson(R × proportions).
Log-likelihood for Poisson model (matrix version)
}
\keyword{internal}
