% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_loglikelihood_vector}
\alias{poisson_loglikelihood_vector}
\title{Log-likelihood for Poisson model (vector version)}
\usage{
poisson_loglikelihood_vector(propvec, yvec, rvec)
}
\arguments{
\item{propvec}{Proportion vector}

\item{yvec}{Observed counts vector}

\item{rvec}{Total counts vector}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood for Poisson model (vector version)
}
\keyword{internal}
