% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_resolve_exceptions}
\alias{poisson_resolve_exceptions}
\title{Resolve exceptions for Poisson model}
\usage{
poisson_resolve_exceptions(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Resolve exceptions for Poisson model
}
\keyword{internal}
