% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data for SNP-Slice}
\usage{
preprocess_data(data, model)
}
\arguments{
\item{data}{Input data}

\item{model}{Model type}
}
\value{
Processed data list with y, r, and metadata
}
\description{
Preprocess data for SNP-Slice
}
\keyword{internal}
