% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.ess_result}
\alias{print.ess_result}
\title{Print effective sample size results}
\usage{
\method{print}{ess_result}(x, digits = 2, ...)
}
\arguments{
\item{x}{ESS result object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments}
}
\value{
Formatted ESS results
}
\description{
Print effective sample size results
}
