% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{read_snp_data}
\alias{read_snp_data}
\title{Read SNP data from file}
\usage{
read_snp_data(file_path, format = "auto")
}
\arguments{
\item{file_path}{Path to data file}

\item{format}{Data format ("auto", "read_counts", "categorical")}
}
\value{
Processed data list
}
\description{
Read SNP data from file
}
\keyword{internal}
