% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates from strain patterns}
\usage{
remove_duplicates(dd)
}
\arguments{
\item{dd}{Binary matrix of strain patterns}
}
\value{
List with assignments and unique dictionary
}
\description{
Remove duplicates from strain patterns
}
\keyword{internal}
