% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC for SNP-Slice}
\usage{
run_mcmc(model_obj, n_sample, n_burnin, gap, verbose, store_mcmc)
}
\arguments{
\item{model_obj}{Model object}

\item{n_sample}{Number of post-burn-in iterations to retain}

\item{n_burnin}{Number of iterations discarded before sampling begins}

\item{gap}{Early stopping threshold}

\item{verbose}{Whether to print progress}

\item{store_mcmc}{Whether to store full MCMC samples}
}
\value{
MCMC results
}
\description{
Run MCMC for SNP-Slice
}
\details{
Only sampling iterations are returned when \code{store_mcmc = TRUE}.
}
\keyword{internal}
