% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{sample_index_logweights}
\alias{sample_index_logweights}
\title{Sample from categorical distribution with log weights}
\usage{
sample_index_logweights(lw, xgrid)
}
\arguments{
\item{lw}{Log weights}

\item{xgrid}{Grid of values}
}
\value{
Sampled value
}
\description{
Sample from categorical distribution with log weights
}
\keyword{internal}
