% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{slice_init}
\alias{slice_init}
\title{Initialize slice sampler}
\usage{
slice_init(state, model_obj)
}
\arguments{
\item{state}{Initial state}

\item{model_obj}{Model object}
}
\value{
State with slice sampler initialized
}
\description{
Initialize slice sampler
}
\keyword{internal}
