% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_iter}
\alias{slice_iter}
\title{Run one slice sampling iteration}
\usage{
slice_iter(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Run one slice sampling iteration
}
\keyword{internal}
