% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_a}
\alias{slice_update_a}
\title{Update allocation matrix A}
\usage{
slice_update_a(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update allocation matrix A
}
\keyword{internal}
