% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_a_r}
\alias{slice_update_a_r}
\title{R reference implementation of allocation-matrix updates}
\usage{
slice_update_a_r(state, model_obj)
}
\description{
R reference implementation of allocation-matrix updates
}
\keyword{internal}
