% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_d}
\alias{slice_update_d}
\title{Update dictionary matrix D}
\usage{
slice_update_d(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update dictionary matrix D
}
\keyword{internal}
