% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_mu}
\alias{slice_update_mu}
\title{Update stick-breaking weights mu}
\usage{
slice_update_mu(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update stick-breaking weights mu
}
\keyword{internal}
