% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_mu_r}
\alias{slice_update_mu_r}
\title{R reference implementation of stick-breaking weight updates}
\usage{
slice_update_mu_r(state, model_obj)
}
\description{
R reference implementation of stick-breaking weight updates
}
\keyword{internal}
