% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_s}
\alias{slice_update_s}
\title{Update slice variable s}
\usage{
slice_update_s(state, model_obj)
}
\arguments{
\item{state}{Current state}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update slice variable s
}
\keyword{internal}
