% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_s_r}
\alias{slice_update_s_r}
\title{R reference implementation of slice-variable update}
\usage{
slice_update_s_r(state, model_obj)
}
\description{
R reference implementation of slice-variable update
}
\keyword{internal}
