% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\name{start_timer}
\alias{start_timer}
\title{Start timing a function}
\usage{
start_timer(function_name)
}
\arguments{
\item{function_name}{Name of the function being timed}
}
\description{
Start timing a function
}
\keyword{internal}
