% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate algorithm parameters}
\usage{
validate_parameters(alpha, rho, threshold)
}
\arguments{
\item{alpha}{IBP concentration parameter}

\item{rho}{Dictionary sparsity parameter}

\item{threshold}{Single infection threshold}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate algorithm parameters
}
\keyword{internal}
