% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{calculate_allele_counts_single}
\alias{calculate_allele_counts_single}
\title{Calculate allele counts for a single sample}
\usage{
calculate_allele_counts_single(
  A,
  D,
  snp_indices,
  r0_values,
  r1_values,
  sep = "|"
)
}
\description{
Calculate allele counts for a single sample
}
\keyword{internal}
