% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_loglikelihood_matrix}
\alias{categorical_loglikelihood_matrix}
\title{Categorical Model for SNP-Slice}
\usage{
categorical_loglikelihood_matrix(A, D, model_obj)
}
\arguments{
\item{A}{Allocation matrix}

\item{D}{Dictionary matrix}

\item{model_obj}{Model object containing data}
}
\value{
Log-likelihood value
}
\description{
Implementation of the categorical observation model for SNP-Slice.
This model handles categorical observations (0, 0.5, 1) with error parameters.
Log-likelihood for categorical model (matrix version)
}
\keyword{internal}
