% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_loglikelihood_vector}
\alias{categorical_loglikelihood_vector}
\title{Log-likelihood for categorical model (vector version)}
\usage{
categorical_loglikelihood_vector(propvec, yvec, rvec = NULL)
}
\arguments{
\item{propvec}{Proportion vector}

\item{yvec}{Observed categorical values}

\item{rvec}{Not used for categorical model (kept for interface consistency)}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood for categorical model (vector version)
}
\keyword{internal}
