% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{create_model}
\alias{create_model}
\title{Create model object}
\usage{
create_model(model, processed_data, alpha = 2.6, rho = 0.5, ...)
}
\arguments{
\item{model}{Model type}

\item{processed_data}{Processed data}

\item{...}{Additional model parameters}
}
\value{
Model object with initialization and likelihood functions
}
\description{
Create model object
}
\keyword{internal}
