% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_allocations}
\alias{extract_allocations}
\title{Extract allocation information from SNP-Slice results}
\usage{
extract_allocations(results)
}
\arguments{
\item{results}{SNP-Slice results object}
}
\value{
List containing allocation information
}
\description{
Extract allocation information from SNP-Slice results
}
