% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_strains}
\alias{extract_strains}
\title{Extract strain information from SNP-Slice results}
\usage{
extract_strains(results)
}
\arguments{
\item{results}{SNP-Slice results object}
}
\value{
List containing strain information
}
\description{
Extract strain information from SNP-Slice results
}
