% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_logger.R
\docType{data}
\name{performance_env}
\alias{performance_env}
\title{Performance Logger for SNP-Slice}
\format{
An object of class \code{environment} of length 0.
}
\usage{
performance_env
}
\description{
Utility functions for tracking performance and identifying bottlenecks
}
\keyword{datasets}
