% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot convergence diagnostics}
\usage{
plot_convergence(results, type = "logpost")
}
\arguments{
\item{results}{SNP-Slice results object}

\item{type}{Type of plot ("logpost", "kstar", "n_strains")}
}
\value{
Plot object (if ggplot2 is available)
}
\description{
Plot convergence diagnostics
}
