% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{refresh_feature}
\alias{refresh_feature}
\title{Refresh feature with random dictionary}
\usage{
refresh_feature(state, k, rho, P)
}
\arguments{
\item{state}{Current state}

\item{k}{Feature index}

\item{rho}{Dictionary sparsity parameter}

\item{P}{Number of SNPs}
}
\value{
Updated state
}
\description{
Refresh feature with random dictionary
}
\keyword{internal}
