% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC for SNP-Slice}
\usage{
run_mcmc(model_obj, n_mcmc, burnin, gap, verbose, store_mcmc)
}
\arguments{
\item{model_obj}{Model object}

\item{n_mcmc}{Number of MCMC iterations}

\item{burnin}{Burn-in period}

\item{gap}{Early stopping threshold}

\item{verbose}{Whether to print progress}

\item{store_mcmc}{Whether to store full MCMC samples}
}
\value{
MCMC results
}
\description{
Run MCMC for SNP-Slice
}
\keyword{internal}
