% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_a_local}
\alias{slice_update_a_local}
\title{Update single element of allocation matrix A}
\usage{
slice_update_a_local(state, i, k, model_obj)
}
\arguments{
\item{state}{Current state}

\item{i}{Host index}

\item{k}{Strain index}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update single element of allocation matrix A
}
\keyword{internal}
