% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_update_d_local}
\alias{slice_update_d_local}
\title{Update single element of dictionary matrix D}
\usage{
slice_update_d_local(state, k, p, model_obj)
}
\arguments{
\item{state}{Current state}

\item{k}{Strain index}

\item{p}{SNP index}

\item{model_obj}{Model object}
}
\value{
Updated state
}
\description{
Update single element of dictionary matrix D
}
\keyword{internal}
