% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary.snp_slice_results}
\alias{summary.snp_slice_results}
\title{Print summary of SNP-Slice results}
\usage{
\method{summary}{snp_slice_results}(results)
}
\arguments{
\item{results}{SNP-Slice results object}
}
\value{
Summary information
}
\description{
Print summary of SNP-Slice results
}
