% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate input data for SNP-Slice}
\usage{
validate_input_data(data, model)
}
\arguments{
\item{data}{Input data to validate}

\item{model}{Model type to validate against}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate input data for SNP-Slice
}
\keyword{internal}
