% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{validate_mcmc_settings}
\alias{validate_mcmc_settings}
\title{Validate MCMC settings}
\usage{
validate_mcmc_settings(n_mcmc, burnin, gap)
}
\arguments{
\item{n_mcmc}{Number of MCMC iterations}

\item{burnin}{Burn-in period}

\item{gap}{Early stopping threshold}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate MCMC settings
}
\keyword{internal}
